AI Navigation / Animations
I got a pretty decent AI Patrol and Pursue mechanic going with adjustable parameters (cone radius, distance, speed etc). Also added a model and animations for the player to be able to cast shadows. This can be adjusted but as a placeholder it works well.

The AI uses Nav Mesh and we can determine the waypoints that the enemy will rotate between. The rotation of where the enemy goes is also "random" so it doesn't just rotate between them in a specific order. The player can leave the sight of the enemy for a brief amount of time to reset the AI. The AI will then go back to patrolling locally.